pyproject.toml
[tool.poetry]
name = "desafio-dev-junior"
version = "0.1.0"
description = ""
authors = ["Jorge Vilaça"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
django-localflavor = "^3.0.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
